requirements.txt
streamlit>=1.31.0
pydantic>=2.6.0
pandas>=2.0.0
pyyaml>=6.0.0
google-genai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.1
